rounds interpolated width

+++ Interpolate/Interpolate.py
@@ -523,7 +523,7 @@ class InterpolateWindow(object):
 				targetPoint = self.allTargetPoints[i][j]
 				gI[i].points[j].x = int((sourcePoint[0] + ((targetPoint[0] - sourcePoint[0]) * valueX/1000)) * self.scaleXValue/100)
 				gI[i].points[j].y = int((sourcePoint[1] + ((targetPoint[1] - sourcePoint[1]) * valueY/1000)) * self.scaleYValue/100)
-				gI.width = (gS.width + ((gT.width - gS.width) * valueX/1000)) * self.scaleXValue/100
+				gI.width = int((gS.width + ((gT.width - gS.width) * valueX/1000)) * self.scaleXValue/100)
 		return gI
 			
 	def collaPolate(self, gS, gT, barIncrement):
